Django/DjangoRestFramework/Todo/requirements.txt
django 
djangorestframework
djangorestframework-simplejwt
django-extensions
pygments
Pillow